Tidy onBindViewHolder
Using a combo of string and position was unnecessary and we can instead just use the animal name which we know to be unique.

We can also use getAdapterPosition for the position (rather than making the parameter final, which you should never do) of item that's used within the AnimalViewPagerFragment.