Fix picasso fragment to fragment return shared element transition
Removing the picasso call to centerCrop the image and relying on the
centerCrop attribute for both ImageViews seems to have fixed this.

Added a BaseFragment with a convenience method for listening to when a
view has been laid out so that we cna avoid jumpy transitions.

Added a listener to the PicassoA fragment to ensure we can do the
return transition properly provided everything is loaded and ready.